+++ Portal/Contract/PortalFactoryContract.php
@@ -5,72 +5,53 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\Portal\Contract;
 
 use Heptacom\HeptaConnect\Core\Portal\Exception\AbstractInstantiationException;
-use Heptacom\HeptaConnect\Core\Portal\Exception\ClassNotFoundOnInstantionException;
 use Heptacom\HeptaConnect\Core\Portal\Exception\InaccessableConstructorOnInstantionException;
-use Heptacom\HeptaConnect\Core\Portal\Exception\UnexpectedClassInheritanceOnInstantionException;
 use Heptacom\HeptaConnect\Core\Portal\Exception\UnexpectedRequiredParameterInConstructorOnInstantionException;
+use Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalExtensionContract;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalType;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PortalExtensionType;
 
 abstract class PortalFactoryContract
 {
     /**
-     * @psalm-param class-string<PortalContract> $class
-     *
      * @throws AbstractInstantiationException
      */
-    public function instantiatePortal(string $class): PortalContract
+    public function instantiatePortal(PortalType $class): PortalContract
     {
-        if (!\class_exists($class)) {
-            throw new ClassNotFoundOnInstantionException($class);
-        }
-
-        if (!\is_a($class, PortalContract::class, true)) {
-            throw new UnexpectedClassInheritanceOnInstantionException($class, PortalContract::class);
-        }
-
-        $reflection = new \ReflectionClass($class);
-
-        if (!$reflection->isInstantiable()) {
-            throw new InaccessableConstructorOnInstantionException($class);
-        }
-
-        $ctor = $reflection->getConstructor();
-
-        if ($ctor instanceof \ReflectionMethod && $ctor->getNumberOfRequiredParameters() > 0) {
-            throw new UnexpectedRequiredParameterInConstructorOnInstantionException($class);
-        }
-
-        return new $class();
+        return $this->instantiateObject($class);
     }
 
     /**
-     * @psalm-param class-string<PortalExtensionContract> $class
-     *
      * @throws AbstractInstantiationException
      */
-    public function instantiatePortalExtension(string $class): PortalExtensionContract
+    public function instantiatePortalExtension(PortalExtensionType $class): PortalExtensionContract
     {
-        if (!\class_exists($class)) {
-            throw new ClassNotFoundOnInstantionException($class);
-        }
+        return $this->instantiateObject($class);
+    }
 
-        if (!\is_a($class, PortalExtensionContract::class, true)) {
-            throw new UnexpectedClassInheritanceOnInstantionException($class, PortalExtensionContract::class);
-        }
-
-        $reflection = new \ReflectionClass($class);
+    /**
+     * @throws AbstractInstantiationException
+     *
+     * @return PortalContract|PortalExtensionContract
+     */
+    private function instantiateObject(ClassStringContract $class): object
+    {
+        $reflection = new \ReflectionClass((string) $class);
 
         if (!$reflection->isInstantiable()) {
-            throw new InaccessableConstructorOnInstantionException($class);
+            throw new InaccessableConstructorOnInstantionException((string) $class);
         }
 
         $ctor = $reflection->getConstructor();
 
         if ($ctor instanceof \ReflectionMethod && $ctor->getNumberOfRequiredParameters() > 0) {
-            throw new UnexpectedRequiredParameterInConstructorOnInstantionException($class);
+            throw new UnexpectedRequiredParameterInConstructorOnInstantionException((string) $class);
         }
 
-        return new $class();
+        $classString = $class->getClassString();
+
+        return new $classString();
     }
 }